move error decorators into own package
