+++ app/decorators/__init__.py
@@ -1,2 +1,2 @@
 '''Init the decorators module'''
-from app.decorators.error_aggregation import handle_error as catcher
+from app.decorators.errors.error_aggregation import handle_error as catcher
