+++ tests/decorators/test_game_exception_handler.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 import azure.functions as func
 
-from app.decorators.hundred_and_ten_error import handle_error
+from app.decorators.errors.hundred_and_ten_error import handle_error
 from app.models import HundredAndTenError
 from tests.helpers import build_request
